app/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
google-api-python-client
google-genai
google_auth_oauthlib
fastapi[standard]
streamlit
mangum
pydantic_settings
boto3
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
cryptography
python-jose[cryptography]
python-multipart